Retain spaces when splitting long lines
When long qualifier values are split on spaces to keep the total line length less than 80 characters,
and the final character on a line is a space, that space is deleted. However, this prevents parsing the
original value - it's not possible to know when joining the lines together whether a space should be
inserted or not. For example:

Lipopolysaccharide export system ATP-binding protein LptB

will appear in the Genbank text as:

/product="Lipopolysaccharide export system ATP-binding
 protein LptB"

and when this is read later and newlines are removed, it will be:

Lipopolysaccharide export system ATP-bindingprotein LptB

This commit retains the final space at the end of the line so that the original value can be
reconstructed unambiguously.